+++ tests/src/widgets/LangSwitcherTest.php
@@ -11,8 +11,8 @@ class LangSwitcherTest extends CmsFrontendTestCase
     public function testWidgetOutput()
     {
         $this->assertSame('<ul class="list-element">
-<li class="lang-element-item lang-element-item--active"><a class="lang-link-item lang-link-item--active" href="/luya/envs/dev/public_html/">English</a></li>
-<li class="lang-element-item"><a class="lang-link-item" href="/luya/envs/dev/public_html/de">Deutsch</a></li>
+<li class="lang-element-item lang-element-item--active"><a class="lang-link-item lang-link-item--active" href="http://localhost/luya/envs/dev/public_html/">English</a></li>
+<li class="lang-element-item"><a class="lang-link-item" href="http://localhost/luya/envs/dev/public_html/de">Deutsch</a></li>
 </ul>', LangSwitcher::widget());
     }
 
@@ -23,8 +23,8 @@ class LangSwitcherTest extends CmsFrontendTestCase
         }]);
         
         $this->assertSame('<ul class="list-element">
-<li class="lang-element-item lang-element-item--active"><a class="lang-link-item lang-link-item--active" href="/luya/envs/dev/public_html/">EN</a></li>
-<li class="lang-element-item"><a class="lang-link-item" href="/luya/envs/dev/public_html/de">DE</a></li>
+<li class="lang-element-item lang-element-item--active"><a class="lang-link-item lang-link-item--active" href="http://localhost/luya/envs/dev/public_html/">EN</a></li>
+<li class="lang-element-item"><a class="lang-link-item" href="http://localhost/luya/envs/dev/public_html/de">DE</a></li>
 </ul>', $out);
     }
     
@@ -38,8 +38,8 @@ class LangSwitcherTest extends CmsFrontendTestCase
         
         
         $this->assertSame('<ul class="list-element">
-<li class="lang-element-item"><a class="lang-link-item" href="/luya/envs/dev/public_html/de">Deutsch</a></li>
-<li class="lang-element-item lang-element-item--active"><a class="lang-link-item lang-link-item--active" href="/luya/envs/dev/public_html/">English</a></li>
+<li class="lang-element-item"><a class="lang-link-item" href="http://localhost/luya/envs/dev/public_html/de">Deutsch</a></li>
+<li class="lang-element-item lang-element-item--active"><a class="lang-link-item lang-link-item--active" href="http://localhost/luya/envs/dev/public_html/">English</a></li>
 </ul>', $out);
     }
 
@@ -47,7 +47,7 @@ class LangSwitcherTest extends CmsFrontendTestCase
     {
         $out = LangSwitcher::widget(['noListTag' => true]);
 
-        $this->assertSame('<li class="lang-element-item lang-element-item--active"><a class="lang-link-item lang-link-item--active" href="/luya/envs/dev/public_html/">English</a></li>
-<li class="lang-element-item"><a class="lang-link-item" href="/luya/envs/dev/public_html/de">Deutsch</a></li>', $out);
+        $this->assertSame('<li class="lang-element-item lang-element-item--active"><a class="lang-link-item lang-link-item--active" href="http://localhost/luya/envs/dev/public_html/">English</a></li>
+<li class="lang-element-item"><a class="lang-link-item" href="http://localhost/luya/envs/dev/public_html/de">Deutsch</a></li>', $out);
     }
 }